+++ src/main/java/com/mageddo/dnsproxyserver/config/entrypoint/EnvConfig.java
@@ -1,0 +1,7 @@
+package com.mageddo.dnsproxyserver.config.entrypoint;
+
+public class EnvConfig {
+  public static EnvConfig fromEnv(){
+    throw new UnsupportedOperationException();
+  }
+}
