+++ src/main/java/com/mageddo/dnsproxyserver/config/entrypoint/JsonConfigs.java
@@ -1,0 +1,12 @@
+package com.mageddo.dnsproxyserver.config.entrypoint;
+
+import java.nio.file.Path;
+
+public class JsonConfigs {
+
+  // fixme missing config v1
+
+  public static ConfigurationV2 loadConfigV2(Path path){
+    throw new UnsupportedOperationException();
+  }
+}
