+++ src/main/java/com/mageddo/dnsproxyserver/quarkus/QuarkusConfig.java
@@ -1,6 +1,6 @@
 package com.mageddo.dnsproxyserver.quarkus;
 
-import com.mageddo.dnsproxyserver.config.Configs;
+import com.mageddo.dnsproxyserver.config.Config;
 import org.xbill.DNS.Resolver;
 import org.xbill.DNS.SimpleResolver;
 
@@ -10,6 +10,6 @@ public class QuarkusConfig {
 
   @Produces
   public Resolver simpleResolver() {
-    return new SimpleResolver(Configs.findRemoverSolverConfig().toSocketAddress());
+    return new SimpleResolver(Config.findRemoverSolverConfig().toSocketAddress());
   }
 }
